Oppdater kotlin til 2.3.0 (#1351)

+++ src/test/kotlin/no/nav/helse/flex/FellesTestOppsett.kt
@@ -192,7 +192,7 @@ abstract class FellesTestOppsett : TestOppsettInterfaces {
             .tilJuridiskVurdering()
 }
 
-infix fun Instant.`should be within seconds of`(pair: Pair<Int, Instant>) = this.shouldBeWithinSecondsOf(pair.first.toInt() to pair.second)
+infix fun Instant.`should be within seconds of`(pair: Pair<Int, Instant>) = this.shouldBeWithinSecondsOf(pair.first to pair.second)
 
 infix fun Instant.shouldBeWithinSecondsOf(pair: Pair<Int, Instant>) {
     val (seconds, other) = pair
